+++ CHANGELOG.md
@@ -6,12 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [master] - Unreleased
-### Changed
-- Bump MSRV from 1.39 to 1.42
 
+## [0.4.0] - 2020-07-13
+### Added
 - Add support for async key lookups by way of a `AsyncDecrypter` type which
   wraps the decryption state machine into a `DecrypterInner` type (with
   synchronous key lookup being done by the now-wrapper type `Decrypter`).
+
+### Changed
+- Bump MSRV from 1.39 to 1.42
 - Replace `Decrypter` with `AsyncDecrypter` type in `AsyncDecrypterStream` to
   support using async key lookup functions when using the decrypter stream.
 
@@ -61,7 +64,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial library development
 
-[master]: https://github.com/saltlick-crypto/saltlick-rs/compare/0.3.0...master
+[master]: https://github.com/saltlick-crypto/saltlick-rs/compare/0.4.0...master
+[0.4.0]: https://github.com/saltlick-crypto/saltlick-rs/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/saltlick-crypto/saltlick-rs/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/saltlick-crypto/saltlick-rs/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/saltlick-crypto/saltlick-rs/tree/0.1.0